Move wallet helpers into module
