+++ index.html
@@ -65,7 +65,7 @@
     <!-- <link rel="manifest" href="./manifest.json" /> -->
   </head>
   <body>
-    <script src="./app.js?v=1455-9" type="module"></script>
+    <script src="./app.js?v=1455-10" type="module"></script>
     <script src="./lib.js" type="module"></script>
     <script src="./crypto.js" type="module"></script>
     <div class="container">
@@ -840,6 +840,7 @@
                 <a class="about-link" href="index.html.txt" target="_blank">index.html</a>
                 <a class="about-link" href="styles.css" target="_blank">styles.css</a>
                 <a class="about-link" href="app.js" target="_blank">app.js</a>
+                <a class="about-link" href="wallet.js" target="_blank">wallet.js</a>
                 <br>
                 <a class="about-link" href="crypto.js" target="_blank">crypto.js</a>
                 <a class="about-link" href="lib.js" target="_blank">lib.js</a>
